added validation to inexistent query in query execution

+++ tests/integration/test_stored_query_execution.py
@@ -24,14 +24,27 @@ class StoredQueryCRUDExecution(TornadoAsyncHTTPTestCase, QueryTestCase):
         }'''
         response = self.fetch(request_uri, method="PUT", body=body)
         # 200,201: dont worry if test fails and tearDown is not called to delete the query
-        self.assertTrue(response.code in (200,201))
+        self.assertTrue(response.code in (200, 201))
 
     def tearDown(self):
         super(TornadoAsyncHTTPTestCase, self).setUp()
-        request_uri = "/_query/{0}".format(self.query_id)
-        response = self.fetch(request_uri, method="DELETE")
-        self.assertEqual(response.code, 204)
+        self._delete_stored_query()
 
+    def _delete_stored_query(self):
+        get_status = self._get_stored_query()
+        if get_status == 404:
+            return get_status
+        elif get_status == 200:
+            request_uri = "/_query/{0}".format(self.query_id)
+            response = self.fetch(request_uri, method="DELETE")
+            return response.code
+        else:
+            self.fail("Unexpected GET status {0}".format(get_status))
+
+    def _get_stored_query(self):
+        request_uri = "/_query/{0}".format(self.query_id)
+        response = self.fetch(request_uri)
+        return response.code
 
     @patch("brainiak.utils.i18n.settings", DEFAULT_LANG="en")
     def test_get_query_result(self, mocked_lang):
@@ -47,3 +60,21 @@ class StoredQueryCRUDExecution(TornadoAsyncHTTPTestCase, QueryTestCase):
         self.assertEqual(response.code, 200)
         result = json.loads(response.body)
         self.assertEqual(expected_response, result)
+
+    @patch("brainiak.utils.i18n.settings", DEFAULT_LANG="en")
+    def test_get_query_result_with_missing_param(self, mocked_lang):
+        request_uri = '/_query/{0}/_result'.format(self.query_id)
+        response = self.fetch(request_uri)
+
+        self.assertEqual(response.code, 400)
+
+    @patch("brainiak.utils.i18n.settings", DEFAULT_LANG="en")
+    def test_get_query_result_with_query_not_found(self, mocked_lang):
+        # DELETE!!
+        status = self._delete_stored_query()
+        self.assertEqual(status, 204)
+
+        request_uri = '/_query/{0}/_result'.format(self.query_id)
+        response = self.fetch(request_uri)
+
+        self.assertEqual(response.code, 404)
